split service handler in GrpcServer
