+++ tests-integration/Cargo.toml
@@ -43,7 +43,7 @@ rstest_reuse = "0.5"
 secrecy = "0.8"
 serde.workspace = true
 serde_json = "1.0"
-servers = { workspace = true }
+servers = { workspace = true, features = ["testing"] }
 session = { workspace = true }
 snafu.workspace = true
 sql = { workspace = true }
